// apps/web/src/api/schemas.ts
export const vendaAnaliticaRowSchema = z
  .object({
    data: z.string(),
    codvendedor: z.union([z.number(), z.string()]).optional(),
    nomevendedor: z.string().optional(),
    codprod: z.union([z.number(), z.string()]),
    decprod: z.string(),
    qtdevendida: z.number(),
    und: z.string(),
    qtdeconvertidavd: z.number(),
    precocustoitem: z.number(),
    valorunit: z.number(),
    total: z.number(),
    codcliente: z.union([z.number(), z.string()]),
    nomecliente: z.string(),
    cepcliente: z.string().optional().nullable(),
    totalprodutos: z.number(),
    statuspedido: z.string(),
    datafec: z.string(),
  })
  .passthrough()
// ...
export type VendaAnaliticaRow = z.infer<typeof vendaAnaliticaRowSchema>